requirements.txt
aspose-words==22.4.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
click==8.0.3
colorama==0.4.4
docx-mailmerge==0.5.0
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
idna==2.10
importlib-metadata==4.10.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.1
requests==2.25.1
sh==1.14.2
six==1.16.0
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.8
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.7.0
